findmytie_backend/requirements.txt
celery
channels
django
djangorestframework
django-celery-results
django-cors-headers
pandas
redis
# python -m pip install -U 'channels[daphne]'
